Error page for posts.

+++ routes/kvittra-routes.js
@@ -58,9 +58,9 @@ router.get("/read-kvitter/:id", async (req, res, next) => {
   }
 
   const { token } = req.cookies;
-  if (token && jwt.verify(toke, process.env.JWTSECRET)) {
+  if (token && jwt.verify(token, process.env.JWTSECRET)) {
     if (id) {
-      const kvitterPost = await KvitterModel.findById(req.params.id);
+      const kvitterPost = await KvitterModel.findById(id);
       res.render("posts/single-post", kvitterPost);
     }
   } else {
@@ -69,24 +69,80 @@ router.get("/read-kvitter/:id", async (req, res, next) => {
 });
 
 router.get("/edit/:id", async (req, res, next) => {
-  const kvitterPost = await KvitterModel.findById(req.params.id);
-  res.render("posts/edit-post", kvitterPost);
+  let id = undefined;
+  try {
+    id = req.params.id;
+  } catch {
+    next();
+  }
+
+  const { token } = req.cookies;
+  if (token && jwt.verify(token, process.env.JWTSECRET)) {
+  }
+  if (id) {
+    const kvitterPost = await KvitterModel.findById(req.params.id);
+    res.render("posts/edit-post", kvitterPost);
+  } else {
+    res.redirect("/unauthorized");
+  }
 });
 
 router.post("/edit/:id", async (req, res, next) => {
-  const updatedPost = req.body;
-  await KvitterModel.findById(req.params.id).updateOne(updatedPost);
-  res.redirect("/");
+  let id = undefined;
+  try {
+    id = req.params.id;
+  } catch {
+    next();
+  }
+
+  const { token } = req.cookies;
+  if (token && jwt.verify(token, process.env.JWTSECRET)) {
+  }
+  if (id) {
+    const updatedPost = req.body;
+    await KvitterModel.findById(req.params.id).updateOne(updatedPost);
+    res.redirect("/");
+  } else {
+    res.redirect("/unauthorized");
+  }
 });
 
 router.get("/delete/:id", async (req, res, next) => {
-  const kvitterPost = await KvitterModel.findById(req.params.id);
-  res.render("posts/delete-post", kvitterPost);
+  let id = undefined;
+  try {
+    id = req.params.id;
+  } catch {
+    next();
+  }
+
+  const { token } = req.cookies;
+  if (token && jwt.verify(token, process.env.JWTSECRET)) {
+  }
+  if (id) {
+    const kvitterPost = await KvitterModel.findById(req.params.id);
+    res.render("posts/delete-post", kvitterPost);
+  } else {
+    res.redirect("/unauthorized");
+  }
 });
 
 router.post("/delete/:id", async (req, res, next) => {
-  await KvitterModel.findById(req.params.id).deleteOne();
-  res.redirect("/");
+  let id = undefined;
+  try {
+    id = req.params.id;
+  } catch {
+    next();
+  }
+
+  const { token } = req.cookies;
+  if (token && jwt.verify(token, process.env.JWTSECRET)) {
+  }
+  if (id) {
+    await KvitterModel.findById(req.params.id).deleteOne();
+    res.redirect("/");
+  } else {
+    res.redirect("/unauthorized");
+  }
 });
 
 module.exports = router;
